+++ web/src/views/HostsEntryView.vue
@@ -1,0 +1,15 @@
+<script setup lang="ts">
+import { computed, defineAsyncComponent } from 'vue'
+
+import { session } from '../session'
+
+const view = computed(() =>
+  session.publicReadOnly
+    ? defineAsyncComponent(() => import('./PublicHostsView.vue'))
+    : defineAsyncComponent(() => import('./HostsView.vue')),
+)
+</script>
+
+<template>
+  <component :is="view" />
+</template>